backend/requirements.txt
fastapi
uvicorn
opencv-python
ultralytics
google-generativeai
python-multipart
python-dotenv
pillow
sqlalchemy